pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "br-doc-ocr"
version = "0.1.0"
description = "Brazilian Document OCR using Vision-Language Models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "BR Doc OCR Team"}
]
keywords = ["ocr", "document", "extraction", "vlm", "brazilian", "cnh", "rg"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    # 4.45+ for Qwen2.5-VL; 5.x supported with runtime workaround for ALL_PARALLEL_STYLES
    "transformers>=4.45.0",
    "peft>=0.10.0",
    "datasets>=2.18.0",
    "accelerate>=0.28.0",
    "bitsandbytes>=0.43.0",
    "qwen-vl-utils>=0.0.4",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "sqlalchemy>=2.0.0",
    "pillow>=10.2.0",
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
    "httpx>=0.27.0",
    "vcrpy>=6.0.0",
]
serve = [
    "fastapi>=0.110.0",
    "uvicorn>=0.28.0",
]

[project.scripts]
br-doc-ocr = "br_doc_ocr.cli.main:app"

[project.urls]
Repository = "https://github.com/your-org/br-doc-ocr"
Documentation = "https://github.com/your-org/br-doc-ocr#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
br_doc_ocr = ["schemas/*.json", "py.typed"]

# Ruff configuration (linting)
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "B007",   # unused loop control variable
    "SIM102", "SIM103", "SIM105", "SIM108", "SIM117", "SIM118",  # simplify (style)
    "UP007",  # use X | Y for union (py312)
]

[tool.ruff.lint.per-file-ignores]
# Allow unused fixture args in tests (pytest uses them by name)
"tests/**/*.py" = ["ARG001", "ARG002"]

[tool.ruff.lint.isort]
known-first-party = ["br_doc_ocr"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MyPy configuration (type checking)
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "transformers.*",
    "peft.*",
    "datasets.*",
    "accelerate.*",
    "bitsandbytes.*",
    "qwen_vl_utils.*",
    "torch.*",
    "PIL.*",
    "vcrpy.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "contract: Contract tests",
    "slow: Slow tests (model loading)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/br_doc_ocr"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
